move functions to include folder
