+++ include/config.py
@@ -1,9 +1,7 @@
-from dags.functions.transform_data import (
-    transform_bcra_from_api,
-    transform_mep_usd_from_criptoya_api,
-    transform_other_usd_from_criptoya_api,
-    transform_usdt_from_criptoya_api,
-)
+from include.transform_data import (transform_bcra_from_api,
+                                    transform_mep_usd_from_criptoya_api,
+                                    transform_other_usd_from_criptoya_api,
+                                    transform_usdt_from_criptoya_api)
 
 TRANSFORM_FUNCTIONS = {
     "mep": {
