+++ tests/functions/test_extract_function.py
@@ -2,8 +2,9 @@ import json
 from unittest.mock import patch
 
 import pytest
-from functions.extract_data import extract_data_from_api
 from requests.exceptions import RequestException
+
+from include.extract_data import extract_data_from_api
 
 
 @pytest.fixture
